Try fixing the tests without optional deps
