+++ tests/auxil/bot_method_checks.py
@@ -19,10 +19,10 @@
 import datetime
 import functools
 import inspect
+import os
 from typing import Any, Callable, Dict, Iterable, List
 
 import pytest
-import pytz
 
 from telegram import (
     Bot,
@@ -38,6 +38,11 @@ from telegram._utils.defaultvalue import DEFAULT_NONE, DefaultValue
 from telegram.constants import InputMediaType
 from telegram.ext import Defaults, ExtBot
 from telegram.request import RequestData
+from tests.auxil.object_conversions import env_var_2_bool
+
+TEST_WITH_OPT_DEPS = env_var_2_bool(os.getenv("TEST_WITH_OPT_DEPS", True))
+if TEST_WITH_OPT_DEPS:
+    import pytz
 
 
 def check_shortcut_signature(
