+++ tests/test_no_passport.py
@@ -32,7 +32,7 @@ import pytest
 
 from telegram import _bot as bot
 from telegram._passport import credentials as credentials
-from tests.conftest import env_var_2_bool
+from tests.auxil.object_conversions import env_var_2_bool
 
 TEST_WITH_OPT_DEPS = env_var_2_bool(os.getenv("TEST_WITH_OPT_DEPS", True))
