+++ tests/test_updater.py
@@ -31,10 +31,10 @@ from telegram._utils.defaultvalue import DEFAULT_NONE
 from telegram.error import InvalidToken, RetryAfter, TelegramError, TimedOut
 from telegram.ext import ExtBot, InvalidCallbackData, Updater
 from telegram.request import HTTPXRequest
+from tests.auxil.object_conversions import env_var_2_bool
 from tests.conftest import (
     DictBot,
     data_file,
-    env_var_2_bool,
     make_bot,
     make_message,
     make_message_update,
